pyproject.toml
[project]
name = "voice-serve"
version = "0.1.0"
description = "VAPI voice call bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "psycopg2-binary>=2.9.9",
]
